Use double precision (will need some reverts later) Remove measurements.h

+++ dogm/include/kernel/particle_to_grid.h
@@ -26,4 +26,4 @@ SOFTWARE.
 #include "dogm.h"
 #include <device_launch_parameters.h>
 
-__global__ void particleToGridKernel(Particle* particle_array, GridCell* grid_cell_array, float* weight_array, int particle_count);
+__global__ void particleToGridKernel(Particle* particle_array, GridCell* grid_cell_array, double* weight_array, int particle_count);
