+++ dogm/src/kernel/particle_to_grid.cu
@@ -39,7 +39,7 @@ __device__ bool is_last_particle(Particle* particle_array, int particle_count, i
 	return i == particle_count - 1 || particle_array[i].grid_cell_idx != particle_array[i + 1].grid_cell_idx;
 }
 
-__global__ void particleToGridKernel(Particle* particle_array, GridCell* grid_cell_array, float* weight_array, int particle_count)
+__global__ void particleToGridKernel(Particle* particle_array, GridCell* grid_cell_array, double* weight_array, int particle_count)
 {
 	const int i = blockIdx.x * blockDim.x + threadIdx.x;
 	if (i < particle_count)
